requirements.txt
# Core
numpy
scipy
pandas
matplotlib
tqdm
pyyaml

# Jupyter
jupyter
ipykernel

# DL
torch
torchaudio

# Audio
librosa>=0.10.0
soundfile

# Logging
tensorboard

# Dataset manager
soundata